Update Radar to match format given in fakeData, had to change tests cases in test_radar to do this
